+++ src/XRoadLib/IParameterNameProvider.cs
@@ -1,0 +1,9 @@
+﻿using System.Reflection;
+
+namespace XRoadLib
+{
+    public interface IParameterNameProvider
+    {
+        string GetParameterName(ParameterInfo parameterContract, ParameterInfo parameterImpl);
+    }
+}